switched scripts functions over to index, completed domUpdate functions, added a few error handling alerts, and got 2 tests passing for domUpdates
