+++ test/activity-test.js
@@ -15,145 +15,145 @@ describe('Activity', function() {
 
   beforeEach(function() {
     activityData = [{
-        "userID": 1,
-        "date": "2019/06/15",
-        "numSteps": 3577,
-        "minutesActive": 140,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/15",
-        "numSteps": 4294,
-        "minutesActive": 138,
-        "flightsOfStairs": 10
-      },
-      {
-        "userID": 3,
-        "date": "2019/06/15",
-        "numSteps": 7402,
-        "minutesActive": 116,
-        "flightsOfStairs": 33
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/15",
-        "numSteps": 3486,
-        "minutesActive": 114,
-        "flightsOfStairs": 32
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/15",
-        "numSteps": 11374,
-        "minutesActive": 213,
-        "flightsOfStairs": 13
-      },
-      {
-        "userID": 6,
-        "date": "2019/06/15",
-        "numSteps": 14810,
-        "minutesActive": 287,
-        "flightsOfStairs": 18
-      },
-      {
-        "userID": 7,
-        "date": "2019/06/15",
-        "numSteps": 2634,
-        "minutesActive": 107,
-        "flightsOfStairs": 5
-      },
-      {
-        "userID": 11,
-        "date": "2019/06/15",
-        "numSteps": 10333,
-        "minutesActive": 114,
-        "flightsOfStairs": 31
-      },
-      {
-        "userID": 11,
-        "date": "2019/06/15",
-        "numSteps": 6389,
-        "minutesActive": 41,
-        "flightsOfStairs": 33
-      },
-      {
-        "userID": 10,
-        "date": "2019/06/15",
-        "numSteps": 8015,
-        "minutesActive": 106,
-        "flightsOfStairs": 37
-      },
-      {
-        "userID": 11,
-        "date": "2019/06/15",
-        "numSteps": 11652,
-        "minutesActive": 20,
-        "flightsOfStairs": 24
-      },
-      {
-        "userID": 12,
-        "date": "2019/06/15",
-        "numSteps": 9256,
-        "minutesActive": 108,
-        "flightsOfStairs": 2
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/16",
-        "numSteps": 5000,
-        "minutesActive": 12,
-        "flightsOfStairs": 14
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/17",
-        "numSteps": 9303,
-        "minutesActive": 45,
-        "flightsOfStairs": 9
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/18",
-        "numSteps": 3000,
-        "minutesActive": 62,
-        "flightsOfStairs": 23
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/19",
-        "numSteps": 9303,
-        "minutesActive": 4,
-        "flightsOfStairs": 2
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/20",
-        "numSteps": 9303,
-        "minutesActive": 7,
-        "flightsOfStairs": 4
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/21",
-        "numSteps": 12000,
-        "minutesActive": 13,
-        "flightsOfStairs": 26
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/22",
-        "numSteps": 9303,
-        "minutesActive": 21,
-        "flightsOfStairs": 14
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/23",
-        "numSteps": 9000,
-        "minutesActive": 8,
-        "flightsOfStairs": 9
-      }
+      "userID": 1,
+      "date": "2019/06/15",
+      "numSteps": 3577,
+      "minutesActive": 140,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/15",
+      "numSteps": 4294,
+      "minutesActive": 138,
+      "flightsOfStairs": 10
+    },
+    {
+      "userID": 3,
+      "date": "2019/06/15",
+      "numSteps": 7402,
+      "minutesActive": 116,
+      "flightsOfStairs": 33
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/15",
+      "numSteps": 3486,
+      "minutesActive": 114,
+      "flightsOfStairs": 32
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/15",
+      "numSteps": 11374,
+      "minutesActive": 213,
+      "flightsOfStairs": 13
+    },
+    {
+      "userID": 6,
+      "date": "2019/06/15",
+      "numSteps": 14810,
+      "minutesActive": 287,
+      "flightsOfStairs": 18
+    },
+    {
+      "userID": 7,
+      "date": "2019/06/15",
+      "numSteps": 2634,
+      "minutesActive": 107,
+      "flightsOfStairs": 5
+    },
+    {
+      "userID": 11,
+      "date": "2019/06/15",
+      "numSteps": 10333,
+      "minutesActive": 114,
+      "flightsOfStairs": 31
+    },
+    {
+      "userID": 11,
+      "date": "2019/06/15",
+      "numSteps": 6389,
+      "minutesActive": 41,
+      "flightsOfStairs": 33
+    },
+    {
+      "userID": 10,
+      "date": "2019/06/15",
+      "numSteps": 8015,
+      "minutesActive": 106,
+      "flightsOfStairs": 37
+    },
+    {
+      "userID": 11,
+      "date": "2019/06/15",
+      "numSteps": 11652,
+      "minutesActive": 20,
+      "flightsOfStairs": 24
+    },
+    {
+      "userID": 12,
+      "date": "2019/06/15",
+      "numSteps": 9256,
+      "minutesActive": 108,
+      "flightsOfStairs": 2
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/16",
+      "numSteps": 5000,
+      "minutesActive": 12,
+      "flightsOfStairs": 14
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/17",
+      "numSteps": 9303,
+      "minutesActive": 45,
+      "flightsOfStairs": 9
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/18",
+      "numSteps": 3000,
+      "minutesActive": 62,
+      "flightsOfStairs": 23
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/19",
+      "numSteps": 9303,
+      "minutesActive": 4,
+      "flightsOfStairs": 2
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/20",
+      "numSteps": 9303,
+      "minutesActive": 7,
+      "flightsOfStairs": 4
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/21",
+      "numSteps": 12000,
+      "minutesActive": 13,
+      "flightsOfStairs": 26
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/22",
+      "numSteps": 9303,
+      "minutesActive": 21,
+      "flightsOfStairs": 14
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/23",
+      "numSteps": 9000,
+      "minutesActive": 8,
+      "flightsOfStairs": 9
+    }
     ];
 
     user1 = new User({
@@ -315,110 +315,110 @@ describe('Friend Activity', function() {
 
   beforeEach(function() {
     activityData = [{
-        "userID": 1,
-        "date": "2019/06/15",
-        "numSteps": 3577,
-        "minutesActive": 140,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/14",
-        "numSteps": 4294,
-        "minutesActive": 138,
-        "flightsOfStairs": 10
-      },
-      {
-        "userID": 3,
-        "date": "2019/06/13",
-        "numSteps": 7402,
-        "minutesActive": 116,
-        "flightsOfStairs": 33
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/12",
-        "numSteps": 3486,
-        "minutesActive": 114,
-        "flightsOfStairs": 32
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/14",
-        "numSteps": 11374,
-        "minutesActive": 213,
-        "flightsOfStairs": 13
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/13",
-        "numSteps": 14810,
-        "minutesActive": 287,
-        "flightsOfStairs": 18
-      },
-      {
-        "userID": 3,
-        "date": "2019/06/12",
-        "numSteps": 2634,
-        "minutesActive": 107,
-        "flightsOfStairs": 5
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/11",
-        "numSteps": 10333,
-        "minutesActive": 114,
-        "flightsOfStairs": 31
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/02",
-        "numSteps": 6389,
-        "minutesActive": 41,
-        "flightsOfStairs": 33
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/03",
-        "numSteps": 8015,
-        "minutesActive": 106,
-        "flightsOfStairs": 37
-      },
-      {
-        "userID": 3,
-        "date": "2019/06/19",
-        "numSteps": 11652,
-        "minutesActive": 20,
-        "flightsOfStairs": 24
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/15",
-        "numSteps": 9256,
-        "minutesActive": 108,
-        "flightsOfStairs": 2
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/16",
-        "numSteps": 3578,
-        "minutesActive": 140,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/17",
-        "numSteps": 3579,
-        "minutesActive": 141,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/18",
-        "numSteps": 3580,
-        "minutesActive": 142,
-        "flightsOfStairs": 16
-      }
+      "userID": 1,
+      "date": "2019/06/15",
+      "numSteps": 3577,
+      "minutesActive": 140,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/14",
+      "numSteps": 4294,
+      "minutesActive": 138,
+      "flightsOfStairs": 10
+    },
+    {
+      "userID": 3,
+      "date": "2019/06/13",
+      "numSteps": 7402,
+      "minutesActive": 116,
+      "flightsOfStairs": 33
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/12",
+      "numSteps": 3486,
+      "minutesActive": 114,
+      "flightsOfStairs": 32
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/14",
+      "numSteps": 11374,
+      "minutesActive": 213,
+      "flightsOfStairs": 13
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/13",
+      "numSteps": 14810,
+      "minutesActive": 287,
+      "flightsOfStairs": 18
+    },
+    {
+      "userID": 3,
+      "date": "2019/06/12",
+      "numSteps": 2634,
+      "minutesActive": 107,
+      "flightsOfStairs": 5
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/11",
+      "numSteps": 10333,
+      "minutesActive": 114,
+      "flightsOfStairs": 31
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/02",
+      "numSteps": 6389,
+      "minutesActive": 41,
+      "flightsOfStairs": 33
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/03",
+      "numSteps": 8015,
+      "minutesActive": 106,
+      "flightsOfStairs": 37
+    },
+    {
+      "userID": 3,
+      "date": "2019/06/19",
+      "numSteps": 11652,
+      "minutesActive": 20,
+      "flightsOfStairs": 24
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/15",
+      "numSteps": 9256,
+      "minutesActive": 108,
+      "flightsOfStairs": 2
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/16",
+      "numSteps": 3578,
+      "minutesActive": 140,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/17",
+      "numSteps": 3579,
+      "minutesActive": 141,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/18",
+      "numSteps": 3580,
+      "minutesActive": 142,
+      "flightsOfStairs": 16
+    }
     ];
 
     activity = new Activity(activityData);
@@ -468,78 +468,78 @@ describe('Friend Activity', function() {
 
   it('should get a users friend lists activity', function() {
     expect(activity.getFriendsActivity(user4, userRepo)).to.eql([{
-        "userID": 1,
-        "date": "2019/06/15",
-        "numSteps": 3577,
-        "minutesActive": 140,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/14",
-        "numSteps": 11374,
-        "minutesActive": 213,
-        "flightsOfStairs": 13
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/02",
-        "numSteps": 6389,
-        "minutesActive": 41,
-        "flightsOfStairs": 33
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/16",
-        "numSteps": 3578,
-        "minutesActive": 140,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/17",
-        "numSteps": 3579,
-        "minutesActive": 141,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 1,
-        "date": "2019/06/18",
-        "numSteps": 3580,
-        "minutesActive": 142,
-        "flightsOfStairs": 16
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/14",
-        "numSteps": 4294,
-        "minutesActive": 138,
-        "flightsOfStairs": 10
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/13",
-        "numSteps": 14810,
-        "minutesActive": 287,
-        "flightsOfStairs": 18
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/03",
-        "numSteps": 8015,
-        "minutesActive": 106,
-        "flightsOfStairs": 37
-      }
+      "userID": 1,
+      "date": "2019/06/15",
+      "numSteps": 3577,
+      "minutesActive": 140,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/14",
+      "numSteps": 11374,
+      "minutesActive": 213,
+      "flightsOfStairs": 13
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/02",
+      "numSteps": 6389,
+      "minutesActive": 41,
+      "flightsOfStairs": 33
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/16",
+      "numSteps": 3578,
+      "minutesActive": 140,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/17",
+      "numSteps": 3579,
+      "minutesActive": 141,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 1,
+      "date": "2019/06/18",
+      "numSteps": 3580,
+      "minutesActive": 142,
+      "flightsOfStairs": 16
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/14",
+      "numSteps": 4294,
+      "minutesActive": 138,
+      "flightsOfStairs": 10
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/13",
+      "numSteps": 14810,
+      "minutesActive": 287,
+      "flightsOfStairs": 18
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/03",
+      "numSteps": 8015,
+      "minutesActive": 106,
+      "flightsOfStairs": 37
+    }
     ]);
   });
 
   it('should get a users ranked friendslist activity for a chosen week', function() {
     expect(activity.getFriendsAverageStepsForWeek(user4, "2019/06/15", userRepo)).to.eql([{
-        '2': 9552
-      },
-      {
-        '1': 7475.5
-      }
+      '2': 9552
+    },
+    {
+      '1': 7475.5
+    }
     ]);
   });
