+++ test/sleep-test.js
@@ -17,221 +17,221 @@ describe('Sleep', function() {
 
   beforeEach(function() {
     sleepData = [{
-        "userID": 1,
-        "date": "2017/06/15",
-        "hoursSlept": 6.1,
-        "sleepQuality": 2.2
-      },
-      {
-        "userID": 2,
-        "date": "2017/06/15",
-        "hoursSlept": 7,
-        "sleepQuality": 4.7
-      },
-      {
-        "userID": 3,
-        "date": "2017/06/15",
-        "hoursSlept": 2,
-        "sleepQuality": 3
-      },
-      {
-        "userID": 4,
-        "date": "2017/06/15",
-        "hoursSlept": 5.4,
-        "sleepQuality": 3
-      },
-      {
-        "userID": 1,
-        "date": "2018/07/15",
-        "hoursSlept": 4.1,
-        "sleepQuality": 3.6
-      },
-      {
-        "userID": 2,
-        "date": "2018/07/15",
-        "hoursSlept": 9.6,
-        "sleepQuality": 2.9
-      },
-      {
-        "userID": 3,
-        "date": "2018/07/15",
-        "hoursSlept": 2,
-        "sleepQuality": 3
-      },
-      {
-        "userID": 4,
-        "date": "2018/07/23",
-        "hoursSlept": 8.1,
-        "sleepQuality": 3.5
-      },
-      {
-        "userID": 1,
-        "date": "2019/05/30",
-        "hoursSlept": 8.9,
-        "sleepQuality": 2.2
-      },
-      {
-        "userID": 2,
-        "date": "2019/05/30",
-        "hoursSlept": 4.4,
-        "sleepQuality": 1.6
-      },
-      {
-        "userID": 3,
-        "date": "2019/05/30",
-        "hoursSlept": 4,
-        "sleepQuality": 1
-      },
-      {
-        "userID": 4,
-        "date": "2019/05/30",
-        "hoursSlept": 8,
-        "sleepQuality": 3.4
-      },
-      {
-        "userID": 1,
-        "date": "2019/08/22",
-        "hoursSlept": 10.1,
-        "sleepQuality": 1.8
-      },
-      {
-        "userID": 2,
-        "date": "2019/08/22",
-        "hoursSlept": 6.9,
-        "sleepQuality": 1.2
-      },
-      {
-        "userID": 3,
-        "date": "2019/08/22",
-        "hoursSlept": 4,
-        "sleepQuality": 1
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/21",
-        "hoursSlept": 6.1,
-        "sleepQuality": 3.5
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/20",
-        "hoursSlept": 4.7,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/19",
-        "hoursSlept": 10.1,
-        "sleepQuality": 1.3
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/18",
-        "hoursSlept": 7.9,
-        "sleepQuality": 1.6
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/17",
-        "hoursSlept": 5.9,
-        "sleepQuality": 1.6
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/16",
-        "hoursSlept": 9.6,
-        "sleepQuality": 1
-      },
-      {
-        "userID": 4,
-        "date": "2019/06/15",
-        "hoursSlept": 9,
-        "sleepQuality": 3.1
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/21",
-        "hoursSlept": 6.1,
-        "sleepQuality": 3.5
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/20",
-        "hoursSlept": 4.7,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/19",
-        "hoursSlept": 10.1,
-        "sleepQuality": 3.3
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/18",
-        "hoursSlept": 7.9,
-        "sleepQuality": 3.6
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/17",
-        "hoursSlept": 5.9,
-        "sleepQuality": 3.6
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/16",
-        "hoursSlept": 9.6,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 2,
-        "date": "2019/06/15",
-        "hoursSlept": 9,
-        "sleepQuality": 3.1
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/21",
-        "hoursSlept": 9,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/20",
-        "hoursSlept": 8,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/19",
-        "hoursSlept": 10,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/18",
-        "hoursSlept": 9,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/17",
-        "hoursSlept": 8,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/16",
-        "hoursSlept": 10,
-        "sleepQuality": 4
-      },
-      {
-        "userID": 5,
-        "date": "2019/06/15",
-        "hoursSlept": 9,
-        "sleepQuality": 4
-      }
+      "userID": 1,
+      "date": "2017/06/15",
+      "hoursSlept": 6.1,
+      "sleepQuality": 2.2
+    },
+    {
+      "userID": 2,
+      "date": "2017/06/15",
+      "hoursSlept": 7,
+      "sleepQuality": 4.7
+    },
+    {
+      "userID": 3,
+      "date": "2017/06/15",
+      "hoursSlept": 2,
+      "sleepQuality": 3
+    },
+    {
+      "userID": 4,
+      "date": "2017/06/15",
+      "hoursSlept": 5.4,
+      "sleepQuality": 3
+    },
+    {
+      "userID": 1,
+      "date": "2018/07/15",
+      "hoursSlept": 4.1,
+      "sleepQuality": 3.6
+    },
+    {
+      "userID": 2,
+      "date": "2018/07/15",
+      "hoursSlept": 9.6,
+      "sleepQuality": 2.9
+    },
+    {
+      "userID": 3,
+      "date": "2018/07/15",
+      "hoursSlept": 2,
+      "sleepQuality": 3
+    },
+    {
+      "userID": 4,
+      "date": "2018/07/23",
+      "hoursSlept": 8.1,
+      "sleepQuality": 3.5
+    },
+    {
+      "userID": 1,
+      "date": "2019/05/30",
+      "hoursSlept": 8.9,
+      "sleepQuality": 2.2
+    },
+    {
+      "userID": 2,
+      "date": "2019/05/30",
+      "hoursSlept": 4.4,
+      "sleepQuality": 1.6
+    },
+    {
+      "userID": 3,
+      "date": "2019/05/30",
+      "hoursSlept": 4,
+      "sleepQuality": 1
+    },
+    {
+      "userID": 4,
+      "date": "2019/05/30",
+      "hoursSlept": 8,
+      "sleepQuality": 3.4
+    },
+    {
+      "userID": 1,
+      "date": "2019/08/22",
+      "hoursSlept": 10.1,
+      "sleepQuality": 1.8
+    },
+    {
+      "userID": 2,
+      "date": "2019/08/22",
+      "hoursSlept": 6.9,
+      "sleepQuality": 1.2
+    },
+    {
+      "userID": 3,
+      "date": "2019/08/22",
+      "hoursSlept": 4,
+      "sleepQuality": 1
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/21",
+      "hoursSlept": 6.1,
+      "sleepQuality": 3.5
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/20",
+      "hoursSlept": 4.7,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/19",
+      "hoursSlept": 10.1,
+      "sleepQuality": 1.3
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/18",
+      "hoursSlept": 7.9,
+      "sleepQuality": 1.6
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/17",
+      "hoursSlept": 5.9,
+      "sleepQuality": 1.6
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/16",
+      "hoursSlept": 9.6,
+      "sleepQuality": 1
+    },
+    {
+      "userID": 4,
+      "date": "2019/06/15",
+      "hoursSlept": 9,
+      "sleepQuality": 3.1
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/21",
+      "hoursSlept": 6.1,
+      "sleepQuality": 3.5
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/20",
+      "hoursSlept": 4.7,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/19",
+      "hoursSlept": 10.1,
+      "sleepQuality": 3.3
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/18",
+      "hoursSlept": 7.9,
+      "sleepQuality": 3.6
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/17",
+      "hoursSlept": 5.9,
+      "sleepQuality": 3.6
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/16",
+      "hoursSlept": 9.6,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 2,
+      "date": "2019/06/15",
+      "hoursSlept": 9,
+      "sleepQuality": 3.1
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/21",
+      "hoursSlept": 9,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/20",
+      "hoursSlept": 8,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/19",
+      "hoursSlept": 10,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/18",
+      "hoursSlept": 9,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/17",
+      "hoursSlept": 8,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/16",
+      "hoursSlept": 10,
+      "sleepQuality": 4
+    },
+    {
+      "userID": 5,
+      "date": "2019/06/15",
+      "hoursSlept": 9,
+      "sleepQuality": 4
+    }
     ];
