+++ test/userrepo-test.js
@@ -140,336 +140,336 @@ describe('User Repo', function() {
       userRepo = new UserRepo(users);
 
       hydrationData = [{
-          "userID": 1,
-          "date": "2019/06/15",
-          "numOunces": 37
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/15",
-          "numOunces": 38
-        },
-        {
-          "userID": 3,
-          "date": "2019/05/09",
-          "numOunces": 1
-        },
-        {
-          "userID": 4,
-          "date": "2019/04/15",
-          "numOunces": 36
-        },
-        {
-          "userID": 2,
-          "date": "2018/10/23",
-          "numOunces": 34
-        },
-        {
-          "userID": 1,
-          "date": "2018/06/16",
-          "numOunces": 39
-        },
-        {
-          "userID": 3,
-          "date": "2018/03/30",
-          "numOunces": 2
-        },
-        {
-          "userID": 4,
-          "date": "2018/02/01",
-          "numOunces": 28
-        },
-        {
-          "userID": 1,
-          "date": "2016/08/22",
-          "numOunces": 30
-        },
-        {
-          "userID": 3,
-          "date": "2016/05/14",
-          "numOunces": 3
-        },
-        {
-          "userID": 2,
-          "date": "2016/04/27",
-          "numOunces": 40
-        },
-        {
-          "userID": 4,
-          "date": "2019/03/15",
-          "numOunces": 35
-        },
-        {
-          "userID": 4,
-          "date": "2019/09/20",
-          "numOunces": 40
-        },
-        {
-          "userID": 4,
-          "date": "2019/09/19",
-          "numOunces": 30
-        },
-        {
-          "userID": 4,
-          "date": "2019/09/18",
-          "numOunces": 40
-        },
-        {
-          "userID": 4,
-          "date": "2019/09/17",
-          "numOunces": 40
-        },
-        {
-          "userID": 4,
-          "date": "2019/09/16",
-          "numOunces": 30
-        },
-        {
-          "userID": 4,
-          "date": "2019/09/15",
-          "numOunces": 30
-        },
-        {
-          "userID": 3,
-          "date": "2019/09/17",
-          "numOunces": 30
-        }
+        "userID": 1,
+        "date": "2019/06/15",
+        "numOunces": 37
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/15",
+        "numOunces": 38
+      },
+      {
+        "userID": 3,
+        "date": "2019/05/09",
+        "numOunces": 1
+      },
+      {
+        "userID": 4,
+        "date": "2019/04/15",
+        "numOunces": 36
+      },
+      {
+        "userID": 2,
+        "date": "2018/10/23",
+        "numOunces": 34
+      },
+      {
+        "userID": 1,
+        "date": "2018/06/16",
+        "numOunces": 39
+      },
+      {
+        "userID": 3,
+        "date": "2018/03/30",
+        "numOunces": 2
+      },
+      {
+        "userID": 4,
+        "date": "2018/02/01",
+        "numOunces": 28
+      },
+      {
+        "userID": 1,
+        "date": "2016/08/22",
+        "numOunces": 30
+      },
+      {
+        "userID": 3,
+        "date": "2016/05/14",
+        "numOunces": 3
+      },
+      {
+        "userID": 2,
+        "date": "2016/04/27",
+        "numOunces": 40
+      },
+      {
+        "userID": 4,
+        "date": "2019/03/15",
+        "numOunces": 35
+      },
+      {
+        "userID": 4,
+        "date": "2019/09/20",
+        "numOunces": 40
+      },
+      {
+        "userID": 4,
+        "date": "2019/09/19",
+        "numOunces": 30
+      },
+      {
+        "userID": 4,
+        "date": "2019/09/18",
+        "numOunces": 40
+      },
+      {
+        "userID": 4,
+        "date": "2019/09/17",
+        "numOunces": 40
+      },
+      {
+        "userID": 4,
+        "date": "2019/09/16",
+        "numOunces": 30
+      },
+      {
+        "userID": 4,
+        "date": "2019/09/15",
+        "numOunces": 30
+      },
+      {
+        "userID": 3,
+        "date": "2019/09/17",
+        "numOunces": 30
+      }
       ]
 
       sleepData = [{
-          "userID": 1,
-          "date": "2017/06/15",
-          "hoursSlept": 6.1,
-          "sleepQuality": 2.2
-        },
-        {
-          "userID": 2,
-          "date": "2017/06/15",
-          "hoursSlept": 7,
-          "sleepQuality": 4.7
-        },
-        {
-          "userID": 3,
-          "date": "2017/06/15",
-          "hoursSlept": 2,
-          "sleepQuality": 3
-        },
-        {
-          "userID": 4,
-          "date": "2017/06/15",
-          "hoursSlept": 5.4,
-          "sleepQuality": 3
-        },
-        {
-          "userID": 1,
-          "date": "2018/07/15",
-          "hoursSlept": 4.1,
-          "sleepQuality": 3.6
-        },
-        {
-          "userID": 2,
-          "date": "2018/07/15",
-          "hoursSlept": 9.6,
-          "sleepQuality": 2.9
-        },
-        {
-          "userID": 3,
-          "date": "2018/07/15",
-          "hoursSlept": 2,
-          "sleepQuality": 3
-        },
-        {
-          "userID": 4,
-          "date": "2018/07/23",
-          "hoursSlept": 8.1,
-          "sleepQuality": 3.5
-        },
-        {
-          "userID": 1,
-          "date": "2019/05/30",
-          "hoursSlept": 8.9,
-          "sleepQuality": 2.2
-        },
-        {
-          "userID": 2,
-          "date": "2019/05/30",
-          "hoursSlept": 4.4,
-          "sleepQuality": 1.6
-        },
-        {
-          "userID": 3,
-          "date": "2019/05/30",
-          "hoursSlept": 4,
-          "sleepQuality": 1
-        },
-        {
-          "userID": 4,
-          "date": "2019/05/30",
-          "hoursSlept": 8,
-          "sleepQuality": 3.4
-        },
-        {
-          "userID": 1,
-          "date": "2019/08/22",
-          "hoursSlept": 10.1,
-          "sleepQuality": 1.8
-        },
-        {
-          "userID": 2,
-          "date": "2019/08/22",
-          "hoursSlept": 6.9,
-          "sleepQuality": 1.2
-        },
-        {
-          "userID": 3,
-          "date": "2019/08/22",
-          "hoursSlept": 4,
-          "sleepQuality": 1
-        },
-        {
-          "userID": 4,
-          "date": "2019/06/21",
-          "hoursSlept": 6.1,
-          "sleepQuality": 3.5
-        },
-        {
-          "userID": 4,
-          "date": "2019/06/20",
-          "hoursSlept": 4.7,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 4,
-          "date": "2019/06/19",
-          "hoursSlept": 10.1,
-          "sleepQuality": 1.3
-        },
-        {
-          "userID": 4,
-          "date": "2019/06/18",
-          "hoursSlept": 7.9,
-          "sleepQuality": 1.6
-        },
-        {
-          "userID": 4,
-          "date": "2019/06/17",
-          "hoursSlept": 5.9,
-          "sleepQuality": 1.6
-        },
-        {
-          "userID": 4,
-          "date": "2019/06/16",
-          "hoursSlept": 9.6,
-          "sleepQuality": 1
-        },
-        {
-          "userID": 4,
-          "date": "2019/06/15",
-          "hoursSlept": 9,
-          "sleepQuality": 3.1
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/21",
-          "hoursSlept": 6.1,
-          "sleepQuality": 3.5
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/20",
-          "hoursSlept": 4.7,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/19",
-          "hoursSlept": 10.1,
-          "sleepQuality": 3.3
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/18",
-          "hoursSlept": 7.9,
-          "sleepQuality": 3.6
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/17",
-          "hoursSlept": 5.9,
-          "sleepQuality": 3.6
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/16",
-          "hoursSlept": 9.6,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 2,
-          "date": "2019/06/15",
-          "hoursSlept": 9,
-          "sleepQuality": 3.1
-        },
-        {
-          "userID": 5,
-          "date": "2019/06/21",
-          "hoursSlept": 9,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 5,
-          "date": "2019/06/20",
-          "hoursSlept": 8,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 5,
-          "date": "2019/06/19",
-          "hoursSlept": 10,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 5,
-          "date": "2019/06/18",
-          "hoursSlept": 9,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 5,
-          "date": "2019/06/17",
-          "hoursSlept": 8,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 5,
-          "date": "2019/06/16",
-          "hoursSlept": 10,
-          "sleepQuality": 4
-        },
-        {
-          "userID": 5,
-          "date": "2019/06/15",
-          "hoursSlept": 9,
-          "sleepQuality": 4
-        }
+        "userID": 1,
+        "date": "2017/06/15",
+        "hoursSlept": 6.1,
+        "sleepQuality": 2.2
+      },
+      {
+        "userID": 2,
+        "date": "2017/06/15",
+        "hoursSlept": 7,
+        "sleepQuality": 4.7
+      },
+      {
+        "userID": 3,
+        "date": "2017/06/15",
+        "hoursSlept": 2,
+        "sleepQuality": 3
+      },
+      {
+        "userID": 4,
+        "date": "2017/06/15",
+        "hoursSlept": 5.4,
+        "sleepQuality": 3
+      },
+      {
+        "userID": 1,
+        "date": "2018/07/15",
+        "hoursSlept": 4.1,
+        "sleepQuality": 3.6
+      },
+      {
+        "userID": 2,
+        "date": "2018/07/15",
+        "hoursSlept": 9.6,
+        "sleepQuality": 2.9
+      },
+      {
+        "userID": 3,
+        "date": "2018/07/15",
+        "hoursSlept": 2,
+        "sleepQuality": 3
+      },
+      {
+        "userID": 4,
+        "date": "2018/07/23",
+        "hoursSlept": 8.1,
+        "sleepQuality": 3.5
+      },
+      {
+        "userID": 1,
+        "date": "2019/05/30",
+        "hoursSlept": 8.9,
+        "sleepQuality": 2.2
+      },
+      {
+        "userID": 2,
+        "date": "2019/05/30",
+        "hoursSlept": 4.4,
+        "sleepQuality": 1.6
+      },
+      {
+        "userID": 3,
+        "date": "2019/05/30",
+        "hoursSlept": 4,
+        "sleepQuality": 1
+      },
+      {
+        "userID": 4,
+        "date": "2019/05/30",
+        "hoursSlept": 8,
+        "sleepQuality": 3.4
+      },
+      {
+        "userID": 1,
+        "date": "2019/08/22",
+        "hoursSlept": 10.1,
+        "sleepQuality": 1.8
+      },
+      {
+        "userID": 2,
+        "date": "2019/08/22",
+        "hoursSlept": 6.9,
+        "sleepQuality": 1.2
+      },
+      {
+        "userID": 3,
+        "date": "2019/08/22",
+        "hoursSlept": 4,
+        "sleepQuality": 1
+      },
+      {
+        "userID": 4,
+        "date": "2019/06/21",
+        "hoursSlept": 6.1,
+        "sleepQuality": 3.5
+      },
+      {
+        "userID": 4,
+        "date": "2019/06/20",
+        "hoursSlept": 4.7,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 4,
+        "date": "2019/06/19",
+        "hoursSlept": 10.1,
+        "sleepQuality": 1.3
+      },
+      {
+        "userID": 4,
+        "date": "2019/06/18",
+        "hoursSlept": 7.9,
+        "sleepQuality": 1.6
+      },
+      {
+        "userID": 4,
+        "date": "2019/06/17",
+        "hoursSlept": 5.9,
+        "sleepQuality": 1.6
+      },
+      {
+        "userID": 4,
+        "date": "2019/06/16",
+        "hoursSlept": 9.6,
+        "sleepQuality": 1
+      },
+      {
+        "userID": 4,
+        "date": "2019/06/15",
+        "hoursSlept": 9,
+        "sleepQuality": 3.1
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/21",
+        "hoursSlept": 6.1,
+        "sleepQuality": 3.5
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/20",
+        "hoursSlept": 4.7,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/19",
+        "hoursSlept": 10.1,
+        "sleepQuality": 3.3
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/18",
+        "hoursSlept": 7.9,
+        "sleepQuality": 3.6
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/17",
+        "hoursSlept": 5.9,
+        "sleepQuality": 3.6
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/16",
+        "hoursSlept": 9.6,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 2,
+        "date": "2019/06/15",
+        "hoursSlept": 9,
+        "sleepQuality": 3.1
+      },
+      {
+        "userID": 5,
+        "date": "2019/06/21",
+        "hoursSlept": 9,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 5,
+        "date": "2019/06/20",
+        "hoursSlept": 8,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 5,
+        "date": "2019/06/19",
+        "hoursSlept": 10,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 5,
+        "date": "2019/06/18",
+        "hoursSlept": 9,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 5,
+        "date": "2019/06/17",
+        "hoursSlept": 8,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 5,
+        "date": "2019/06/16",
+        "hoursSlept": 10,
+        "sleepQuality": 4
+      },
+      {
+        "userID": 5,
+        "date": "2019/06/15",
+        "hoursSlept": 9,
+        "sleepQuality": 4
+      }
       ];
     });
     it('should get a users data from its userID in any data set', function() {
       expect(userRepo.getDataFromUserID(1, hydrationData)).to.eql([{
-          "userID": 1,
-          "date": "2019/06/15",
-          "numOunces": 37
-        },
-        {
-          "userID": 1,
-          "date": "2018/06/16",
-          "numOunces": 39
-        },
-        {
-          "userID": 1,
-          "date": "2016/08/22",
-          "numOunces": 30
-        }
+        "userID": 1,
+        "date": "2019/06/15",
+        "numOunces": 37
+      },
+      {
+        "userID": 1,
+        "date": "2018/06/16",
+        "numOunces": 39
+      },
+      {
+        "userID": 1,
+        "date": "2016/08/22",
+        "numOunces": 30
+      }
       ]);
     });
     it('should get a users most recent date using the app', function() {
